+++ index.html
@@ -48,7 +48,7 @@
     </table>
   </div>
 
-  <script type="module" src="index.js"></script>
+  <script type="module" src="js/index.js"></script>
 </body>
 
 </html>
