+++ js/index.js
@@ -1,0 +1,3 @@
+import { renderApp } from './render.js';
+
+renderApp();
